setup-requirements.txt
# Dependencies for Friend-Lite setup scripts (root wizard + all service init scripts)
rich>=13.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0
ruamel-yaml>=0.18.0
omegaconf>=2.3.0